refactor(theme): migrate and refactor type `GlobalToken` by `@ant-design/cssinjs-utils`
