+++ components/_util/responsiveObserver.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import type { GlobalToken } from '../theme/interface';
+import type { GlobalToken } from '../theme/internal';
 import { useToken } from '../theme/internal';
 
 export type Breakpoint = 'xxl' | 'xl' | 'lg' | 'md' | 'sm' | 'xs';
