+++ components/config-provider/__tests__/theme.test.tsx
@@ -7,7 +7,7 @@ import { Button, InputNumber, Select } from '../..';
 import { resetWarned } from '../../_util/warning';
 import { render } from '../../../tests/utils';
 import theme from '../../theme';
-import type { GlobalToken } from '../../theme';
+import type { GlobalToken } from '../../theme/internal';
 import { useToken } from '../../theme/internal';
 
 const { defaultAlgorithm, darkAlgorithm, compactAlgorithm } = theme;
