+++ components/modal/style/index.ts
@@ -3,8 +3,7 @@ import { unit } from '@ant-design/cssinjs';
 
 import { genFocusStyle, resetComponent } from '../../style';
 import { initFadeMotion, initZoomMotion } from '../../style/motion';
-import type { GlobalToken } from '../../theme';
-import type { AliasToken, FullToken, GenerateStyle } from '../../theme/internal';
+import type { AliasToken, FullToken, GenerateStyle, GlobalToken } from '../../theme/internal';
 import { genStyleHooks, mergeToken } from '../../theme/internal';
 import type { GenStyleFn, TokenWithCommonCls } from '../../theme/util/genComponentStyleHook';
